Added AllTests.xml (for Retry can work) Added AnnotationTransformer Added log to Retry Some tests rebased to BuyOrSellCarInputsTest
